+++ src/Model/src/SecondaryFlightControl.cpp
@@ -116,7 +116,7 @@ std::size_t SecondaryFlightControl::count() const noexcept
 const SecondaryFlightControlData &SecondaryFlightControl::interpolate(int64_t timestamp, TimeVariableData::Access access) const noexcept
 {
     const SecondaryFlightControlData *p1, *p2;
-    const int64_t adjustedTimestamp = qMax(timestamp + d->aircraftInfo.timeOffset, 0LL);
+    const int64_t adjustedTimestamp = qMax(timestamp + d->aircraftInfo.timeOffset, int64_t(0));
 
     if (d->currentTimestamp != adjustedTimestamp || d->currentAccess != access) {
